pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["aperushin <aperushin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
django-environ = "^0.10.0"
psycopg2 = "^2.9.6"
gunicorn = "^21.2.0"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
ansible-vault-win = "^1.0.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
